third_party/blink/web_tests/platform/generic/fast/forms/label/label-selection-expected.txt

Test the selection of label associated with input.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.getSelection().toString() is "Some"
PASS successfullyParsed is true

TEST COMPLETE
